Common functionality refactored to traits
