+++ app/Managers/ServiceManager.php
@@ -8,48 +8,11 @@ use App\Managers\SubscriptionManager;
 use App\Jobs\UploadServiceMedia;
 use App\Events\NewService;
 use Carbon\Carbon;
+use App\Managers\Traits\ServiceTrait;
 
 class ServiceManager extends BaseManager 
 {
-
-	/**
-	 * The service that the manager is working on.
-	 * @var App\Service
-	 */
-	protected $service;
-	/**
-	 * The services that the manager is working on.
-	 * @var Collection
-	 */
-	protected $services;
-
-	/**
-	 * Set the service that the manager will be working on.
-	 * 
-	 * @param  [App\Service 	$service
-	 * @return ServiceManager
-	 */
-	public function forService($service)
-	{
-		if ( !$service instanceof \App\Service ) {
-			$this->setError('Service must be an instance of service.', 500);
-		} else {
-			$this->service = $service;
-		}
-
-		return $this;
-	}
-
-	/**
-	 * Return the services that the manager has got.
-	 * @return collection
-	 */
-	public function services() { return $this->services; }
-	/**
-	 * Return the service that the manager has worked on.
-	 * @return App\Service
-	 */
-	public function service() { return $this->service; }
+	use ServiceTrait;
 
 	/**
 	 * Display a listing of the service based on filtering.
